Refactor AEDT simulation utilities into a class-based package structure
